Backend/utils.py
from PIL import Image
import io
import torchvision.transforms as transforms

image_size = 224  # or your training image size

def transform_image(image_bytes):
    transform = transforms.Compose([
        transforms.Resize((image_size, image_size)),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406],
                             std=[0.229, 0.224, 0.225])
    ])
    image = Image.open(io.BytesIO(image_bytes)).convert('RGB')
    return transform(image).unsqueeze(0)  # add batch dim
