Backend/requirements.txt
flask
torch
torchvision
Pillow
flask-cors
